pyproject.toml
[tool.poetry]
name = "tfl-cycle"
version = "0.1.0"
description = "Data engineering bootcamp final project"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<3.12"
apache-airflow = "2.7.3"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"
lxml = "^4.9.3"
pyarrow = "^14.0.0"
psycopg2-binary = "^2.9.9"
apache-airflow-providers-postgres = "^5.7.1"
google-cloud-storage = "^2.13.0"
apache-airflow-providers-google = "^10.11.0"


[tool.poetry.group.dev.dependencies]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
